requirements.txt
Django==3.2.3
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
Pillow==8.2.0